+++ src/MixedModels.jl
@@ -22,7 +22,7 @@ using Tables
 
 using LinearAlgebra: BlasFloat, BlasReal, HermOrSym, PosDefException, copytri!
 using Base: Ryu
-using GLM: Link, canonicallink
+using GLM: Link, canonicallink, linkfun, linkinv
 using StatsModels: TableRegressionModel
 
 using StatsFuns: log2π, normccdf
@@ -85,6 +85,7 @@ export @formula,
        fit,
        fit!,
        fitted,
+       fitted!,
        fixef,
        fixefnames,
        formula,
@@ -124,6 +125,7 @@ export @formula,
        shortestcovint,
        sdest,
        setθ!,
+       simulate,
        simulate!,
        sparse,
        sparseL,
@@ -175,6 +177,7 @@ include("linalg/rankUpdate.jl")
 include("linalg/logdet.jl")
 include("linalg.jl")
 include("simulate.jl")
+include("predict.jl")
 include("bootstrap.jl")
 include("blockdescription.jl")
 include("grouping.jl")